MAPPS-80: map camera moves only when changing route selection, it does not move when refreshing results
